requirements.txt
dj-database-url==0.4.2
Django==1.11.5
djangorestframework==3.6.4
gunicorn==19.7.1
model-mommy==1.4.0
psycopg2==2.7.3.1
pytz==2017.2
selenium==3.5.0
six==1.11.0
whitenoise==3.3.0
